@@ -1,3 +1,3 @@
 module go.opentelemetry.io/contrib
 
-go 1.14
+go 1.15
