@@ -1,13 +1,12 @@
 module go.opentelemetry.io/otel
 
-go 1.14
+go 1.15
 
 require (
-	github.com/google/go-cmp v0.5.4
+	github.com/google/go-cmp v0.5.6
 	github.com/stretchr/testify v1.7.0
-	go.opentelemetry.io/otel/metric v0.17.0
-	go.opentelemetry.io/otel/oteltest v0.17.0
-	go.opentelemetry.io/otel/trace v0.17.0
+	go.opentelemetry.io/otel/oteltest v1.0.0-RC1
+	go.opentelemetry.io/otel/trace v1.0.0-RC1
 )
 
 replace go.opentelemetry.io/otel => ./
@@ -30,19 +29,17 @@ replace go.opentelemetry.io/otel/example/prometheus => ./example/prometheus
 
 replace go.opentelemetry.io/otel/example/zipkin => ./example/zipkin
 
-replace go.opentelemetry.io/otel/exporters/metric/prometheus => ./exporters/metric/prometheus
+replace go.opentelemetry.io/otel/exporters/prometheus => ./exporters/prometheus
 
-replace go.opentelemetry.io/otel/exporters/otlp => ./exporters/otlp
+replace go.opentelemetry.io/otel/exporters/jaeger => ./exporters/jaeger
 
-replace go.opentelemetry.io/otel/exporters/stdout => ./exporters/stdout
-
-replace go.opentelemetry.io/otel/exporters/trace/jaeger => ./exporters/trace/jaeger
-
-replace go.opentelemetry.io/otel/exporters/trace/zipkin => ./exporters/trace/zipkin
+replace go.opentelemetry.io/otel/exporters/zipkin => ./exporters/zipkin
 
 replace go.opentelemetry.io/otel/internal/tools => ./internal/tools
 
 replace go.opentelemetry.io/otel/sdk => ./sdk
+
+replace go.opentelemetry.io/otel/internal/metric => ./internal/metric
 
 replace go.opentelemetry.io/otel/metric => ./metric
 
@@ -53,3 +50,25 @@ replace go.opentelemetry.io/otel/sdk/export/metric => ./sdk/export/metric
 replace go.opentelemetry.io/otel/sdk/metric => ./sdk/metric
 
 replace go.opentelemetry.io/otel/trace => ./trace
+
+replace go.opentelemetry.io/otel/example/passthrough => ./example/passthrough
+
+replace go.opentelemetry.io/otel/exporters/otlp/otlptrace => ./exporters/otlp/otlptrace
+
+replace go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracegrpc => ./exporters/otlp/otlptrace/otlptracegrpc
+
+replace go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracehttp => ./exporters/otlp/otlptrace/otlptracehttp
+
+replace go.opentelemetry.io/otel/exporters/metric/prometheus => ./exporters/metric/prometheus
+
+replace go.opentelemetry.io/otel/exporters/trace/jaeger => ./exporters/trace/jaeger
+
+replace go.opentelemetry.io/otel/exporters/trace/zipkin => ./exporters/trace/zipkin
+
+replace go.opentelemetry.io/otel/exporters/otlp/otlpmetric => ./exporters/otlp/otlpmetric
+
+replace go.opentelemetry.io/otel/exporters/otlp/otlpmetric/otlpmetricgrpc => ./exporters/otlp/otlpmetric/otlpmetricgrpc
+
+replace go.opentelemetry.io/otel/exporters/stdout/stdoutmetric => ./exporters/stdout/stdoutmetric
+
+replace go.opentelemetry.io/otel/exporters/stdout/stdouttrace => ./exporters/stdout/stdouttrace
