@@ -1,7 +1,9 @@
 package commands
 
 import (
+	"context"
 	"fmt"
+	"time"
 
 	"github.com/ardanlabs/service/business/data/schema"
 	"github.com/ardanlabs/service/foundation/database"
@@ -16,7 +18,10 @@ func Seed(cfg database.Config) error {
 	}
 	defer db.Close()
 
-	if err := schema.Seed(db); err != nil {
+	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
+	defer cancel()
+
+	if err := schema.Seed(ctx, db); err != nil {
 		return errors.Wrap(err, "seed database")
 	}
 
