@@ -3,19 +3,19 @@ package commands
 import (
 	"context"
 	"fmt"
-	"log"
 	"time"
 
-	"github.com/ardanlabs/service/business/auth"
 	"github.com/ardanlabs/service/business/data/user"
+	"github.com/ardanlabs/service/business/sys/auth"
 	"github.com/ardanlabs/service/foundation/database"
 	"github.com/pkg/errors"
+	"go.uber.org/zap"
 )
 
 // UserAdd adds new users into the database.
-func UserAdd(traceID string, log *log.Logger, cfg database.Config, email, password string) error {
-	if email == "" || password == "" {
-		fmt.Println("help: useradd <email> <password>")
+func UserAdd(traceID string, log *zap.SugaredLogger, cfg database.Config, name, email, password string) error {
+	if name == "" || email == "" || password == "" {
+		fmt.Println("help: useradd <name> <email> <password>")
 		return ErrHelp
 	}
 
@@ -28,16 +28,17 @@ func UserAdd(traceID string, log *log.Logger, cfg database.Config, email, passwo
 	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
 	defer cancel()
 
-	u := user.New(log, db)
+	store := user.NewStore(log, db)
 
 	nu := user.NewUser{
+		Name:            name,
 		Email:           email,
 		Password:        password,
 		PasswordConfirm: password,
 		Roles:           []string{auth.RoleAdmin, auth.RoleUser},
 	}
 
-	usr, err := u.Create(ctx, traceID, nu, time.Now())
+	usr, err := store.Create(ctx, traceID, nu, time.Now())
 	if err != nil {
 		return errors.Wrap(err, "create user")
 	}
