@@ -1,81 +1,101 @@
 package main
 
 import (
-	"expvar"           // Register the vars handler.
-	_ "net/http/pprof" // Register the pprof handler.
-
 	"context"
+	"expvar" // Calls init function.
 	"fmt"
-	"log"
 	"net/http"
 	"os"
 	"os/signal"
+	"runtime"
 	"syscall"
 	"time"
 
 	"github.com/ardanlabs/conf"
 	"github.com/ardanlabs/service/app/sales-api/handlers"
-	"github.com/ardanlabs/service/business/auth"
+	"github.com/ardanlabs/service/business/sys/auth"
+	"github.com/ardanlabs/service/business/sys/metrics"
 	"github.com/ardanlabs/service/foundation/database"
 	"github.com/ardanlabs/service/foundation/keystore"
+	"github.com/ardanlabs/service/foundation/logger"
 	"github.com/pkg/errors"
 	"go.opentelemetry.io/otel"
-	"go.opentelemetry.io/otel/exporters/trace/zipkin"
+	"go.opentelemetry.io/otel/attribute"
+	"go.opentelemetry.io/otel/exporters/zipkin"
+	"go.opentelemetry.io/otel/sdk/resource"
 	"go.opentelemetry.io/otel/sdk/trace"
+	semconv "go.opentelemetry.io/otel/semconv/v1.4.0"
+	"go.uber.org/automaxprocs/maxprocs"
+	"go.uber.org/zap"
 )
 
 /*
 Need to figure out timeouts for http service.
-You might want to reset your DB_HOST env var during test tear down.
-Service should start even without a DB running yet.
-symbols in profiles: https://github.com/golang/go/issues/23376 / https://github.com/google/pprof/pull/366
 */
 
 // build is the git version of this program. It is set using build flags in the makefile.
 var build = "develop"
 
 func main() {
-	log := log.New(os.Stdout, "SALES : ", log.LstdFlags|log.Lmicroseconds|log.Lshortfile)
 
+	// Construct the application logger.
+	log := logger.New("SALES-API")
+	defer log.Sync()
+
+	// Make sure the program is using the correct
+	// number of threads if a CPU quota is set.
+	if _, err := maxprocs.Set(); err != nil {
+		log.Errorw("startup", zap.Error(err))
+		os.Exit(1)
+	}
+	log.Infow("startup", "GOMAXPROCS", runtime.GOMAXPROCS(0))
+
+	// Perform the startup and shutdown sequence.
 	if err := run(log); err != nil {
-		log.Println("main: error:", err)
+		log.Errorw("startup", "ERROR", err)
 		os.Exit(1)
 	}
 }
 
-func run(log *log.Logger) error {
+func run(log *zap.SugaredLogger) error {
 
 	// =========================================================================
 	// Configuration
 
-	var cfg struct {
+	cfg := struct {
 		conf.Version
 		Web struct {
 			APIHost         string        `conf:"default:0.0.0.0:3000"`
 			DebugHost       string        `conf:"default:0.0.0.0:4000"`
 			ReadTimeout     time.Duration `conf:"default:5s"`
-			WriteTimeout    time.Duration `conf:"default:5s"`
-			ShutdownTimeout time.Duration `conf:"default:5s"`
-		}
-		DB struct {
-			User       string `conf:"default:postgres"`
-			Password   string `conf:"default:postgres,noprint"`
-			Host       string `conf:"default:db"`
-			Name       string `conf:"default:postgres"`
-			DisableTLS bool   `conf:"default:true"`
+			WriteTimeout    time.Duration `conf:"default:10s"`
+			IdleTimeout     time.Duration `conf:"default:120s"`
+			ShutdownTimeout time.Duration `conf:"default:20s"`
 		}
 		Auth struct {
 			KeysFolder string `conf:"default:zarf/keys/"`
 			Algorithm  string `conf:"default:RS256"`
 		}
+		DB struct {
+			User         string `conf:"default:postgres"`
+			Password     string `conf:"default:postgres,mask"`
+			Host         string `conf:"default:db"`
+			Name         string `conf:"default:postgres"`
+			MaxIdleConns int    `conf:"default:0"`
+			MaxOpenConns int    `conf:"default:0"`
+			DisableTLS   bool   `conf:"default:true"`
+		}
 		Zipkin struct {
-			ReporterURI string  `conf:"default:http://zipkin:9411/api/v2/spans"`
+			ReporterURI string  `conf:"default:http://localhost:9411/api/v2/spans"`
 			ServiceName string  `conf:"default:sales-api"`
 			Probability float64 `conf:"default:0.05"`
 		}
+	}{
+		Version: conf.Version{
+			SVN:  build,
+			Desc: "copyright information here",
+		},
 	}
-	cfg.Version.SVN = build
-	cfg.Version.Desc = "copyright information here"
 
 	if err := conf.Parse(os.Args[1:], "SALES", &cfg); err != nil {
 		switch err {
@@ -101,19 +121,19 @@ func run(log *log.Logger) error {
 	// App Starting
 
 	expvar.NewString("build").Set(build)
-	log.Printf("main: Started: Application initializing: version %q", build)
-	defer log.Println("main: Completed")
+	log.Infow("starting service", "version", build)
+	defer log.Infow("shutdown complete")
 
 	out, err := conf.String(&cfg)
 	if err != nil {
 		return errors.Wrap(err, "generating config for output")
 	}
-	log.Printf("main: Config:\n%v\n", out)
+	log.Infow("startup", "config", out)
 
 	// =========================================================================
 	// Initialize authentication support
 
-	log.Println("main: Started: Initializing authentication support")
+	log.Infow("startup", "status", "initializing authentication support")
 
 	// Construct a key store based on the key files stored in
 	// the specified directory.
@@ -130,20 +150,22 @@ func run(log *log.Logger) error {
 	// =========================================================================
 	// Start Database
 
-	log.Println("main: Initializing database support")
+	log.Infow("startup", "status", "initializing database support", "host", cfg.DB.Host)
 
 	db, err := database.Open(database.Config{
-		User:       cfg.DB.User,
-		Password:   cfg.DB.Password,
-		Host:       cfg.DB.Host,
-		Name:       cfg.DB.Name,
-		DisableTLS: cfg.DB.DisableTLS,
+		User:         cfg.DB.User,
+		Password:     cfg.DB.Password,
+		Host:         cfg.DB.Host,
+		Name:         cfg.DB.Name,
+		MaxIdleConns: cfg.DB.MaxIdleConns,
+		MaxOpenConns: cfg.DB.MaxOpenConns,
+		DisableTLS:   cfg.DB.DisableTLS,
 	})
 	if err != nil {
 		return errors.Wrap(err, "connecting to db")
 	}
 	defer func() {
-		log.Printf("main: Database Stopping: %s", cfg.DB.Host)
+		log.Infow("shutdown", "status", "stopping database support", "host", cfg.DB.Host)
 		db.Close()
 	}()
 
@@ -154,69 +176,91 @@ func run(log *log.Logger) error {
 	// compatible with your project. Please review the documentation for
 	// opentelemetry.
 
-	log.Println("main: Initializing OT/Zipkin tracing support")
+	log.Infow("startup", "status", "initializing OT/Zipkin tracing support")
 
-	exporter, err := zipkin.NewRawExporter(
+	exporter, err := zipkin.New(
 		cfg.Zipkin.ReporterURI,
-		cfg.Zipkin.ServiceName,
-		zipkin.WithLogger(log),
+		// zipkin.WithLogger(zap.NewStdLog(log)),
 	)
 	if err != nil {
 		return errors.Wrap(err, "creating new exporter")
 	}
 
-	tp := trace.NewTracerProvider(
-		trace.WithConfig(trace.Config{DefaultSampler: trace.TraceIDRatioBased(cfg.Zipkin.Probability)}),
+	traceProvider := trace.NewTracerProvider(
+		trace.WithSampler(trace.TraceIDRatioBased(cfg.Zipkin.Probability)),
 		trace.WithBatcher(exporter,
 			trace.WithMaxExportBatchSize(trace.DefaultMaxExportBatchSize),
 			trace.WithBatchTimeout(trace.DefaultBatchTimeout),
 			trace.WithMaxExportBatchSize(trace.DefaultMaxExportBatchSize),
 		),
+		trace.WithResource(
+			resource.NewWithAttributes(
+				semconv.SchemaURL,
+				semconv.ServiceNameKey.String(cfg.Zipkin.ServiceName),
+				attribute.String("exporter", "zipkin"),
+			),
+		),
 	)
 
-	otel.SetTracerProvider(tp)
+	// I can only get this working properly using the singleton :(
+	otel.SetTracerProvider(traceProvider)
+	defer traceProvider.Shutdown(context.Background())
 
 	// =========================================================================
 	// Start Debug Service
-	//
-	// /debug/pprof - Added to the default mux by importing the net/http/pprof package.
-	// /debug/vars - Added to the default mux by importing the expvar package.
-	//
-	// Not concerned with shutting this down when the application is shutdown.
 
-	log.Println("main: Initializing debugging support")
+	log.Infow("startup", "status", "debug router started", "host", cfg.Web.DebugHost)
 
+	// The Debug function returns a mux to listen and serve on for all the debug
+	// related endpoints. This include the standard library endpoints.
+
+	// Construct the mux for the debug calls.
+	debugMux := handlers.DebugMux(build, log, db)
+
+	// Start the service listening for debug requests.
+	// Not concerned with shutting this down with load shedding.
 	go func() {
-		log.Printf("main: Debug Listening %s", cfg.Web.DebugHost)
-		if err := http.ListenAndServe(cfg.Web.DebugHost, http.DefaultServeMux); err != nil {
-			log.Printf("main: Debug Listener closed: %v", err)
+		if err := http.ListenAndServe(cfg.Web.DebugHost, debugMux); err != nil {
+			log.Errorw("shutdown", "status", "debug router closed", "host", cfg.Web.DebugHost, "ERROR", err)
 		}
 	}()
 
 	// =========================================================================
 	// Start API Service
 
-	log.Println("main: Initializing API support")
+	log.Infow("startup", "status", "initializing API support")
 
 	// Make a channel to listen for an interrupt or terminate signal from the OS.
 	// Use a buffered channel because the signal package requires it.
 	shutdown := make(chan os.Signal, 1)
-	signal.Notify(shutdown, os.Interrupt, syscall.SIGTERM)
+	signal.Notify(shutdown, syscall.SIGINT, syscall.SIGTERM)
 
+	// Construct the mux for the API calls.
+	apiMux := handlers.APIMux(handlers.APIMuxConfig{
+		Shutdown: shutdown,
+		Log:      log,
+		Metrics:  metrics.New(),
+		Auth:     auth,
+		DB:       db,
+	})
+
+	// Construct a server to service the requests against the mux.
 	api := http.Server{
 		Addr:         cfg.Web.APIHost,
-		Handler:      handlers.API(build, shutdown, log, auth, db),
+		Handler:      apiMux,
 		ReadTimeout:  cfg.Web.ReadTimeout,
 		WriteTimeout: cfg.Web.WriteTimeout,
+		IdleTimeout:  cfg.Web.IdleTimeout,
+		ErrorLog:     zap.NewStdLog(log.Desugar()),
 	}
 
 	// Make a channel to listen for errors coming from the listener. Use a
 	// buffered channel so the goroutine can exit if we don't collect this error.
 	serverErrors := make(chan error, 1)
 
-	// Start the service listening for requests.
+	// Start the service listening for api requests.
 	go func() {
-		log.Printf("main: API listening on %s", api.Addr)
+		log.Infow("startup", "status", "api router started", "host", api.Addr)
 		serverErrors <- api.ListenAndServe()
 	}()
 
@@ -229,7 +273,8 @@ func run(log *log.Logger) error {
 		return errors.Wrap(err, "server error")
 
 	case sig := <-shutdown:
-		log.Printf("main: %v: Start shutdown", sig)
+		log.Infow("shutdown", "status", "shutdown started", "signal", sig)
+		defer log.Infow("shutdown", "status", "shutdown complete", "signal", sig)
 
 		// Give outstanding requests a deadline for completion.
 		ctx, cancel := context.WithTimeout(context.Background(), cfg.Web.ShutdownTimeout)
