@@ -2,63 +2,79 @@ package main
 
 import (
 	"fmt"
-	"log"
 	"net/http"
 	_ "net/http/pprof"
 	"os"
 	"os/signal"
+	"runtime"
 	"syscall"
 	"time"
 
 	"github.com/ardanlabs/conf"
+	"github.com/ardanlabs/service/app/sales-api/handlers"
 	"github.com/ardanlabs/service/app/sidecar/metrics/collector"
 	"github.com/ardanlabs/service/app/sidecar/metrics/publisher"
 	"github.com/ardanlabs/service/app/sidecar/metrics/publisher/expvar"
+	"github.com/ardanlabs/service/foundation/logger"
 	"github.com/pkg/errors"
+	"go.uber.org/automaxprocs/maxprocs"
+	"go.uber.org/zap"
 )
 
 // build is the git version of this program. It is set using build flags in the makefile.
 var build = "develop"
 
 func main() {
-	log := log.New(os.Stdout, "METRICS : ", log.LstdFlags|log.Lmicroseconds|log.Lshortfile)
+
+	// Construct the application logger.
+	log := logger.New("METRICS")
+	defer log.Sync()
+
+	// Make sure the program is using the correct
+	// number of threads if a CPU quota is set.
+	if _, err := maxprocs.Set(); err != nil {
+		log.Errorw("startup", zap.Error(err))
+		os.Exit(1)
+	}
+	log.Infow("startup", "GOMAXPROCS", runtime.GOMAXPROCS(0))
 
 	if err := run(log); err != nil {
-		log.Println("main: error:", err)
+		log.Errorw("startup", "ERROR", err)
 		os.Exit(1)
 	}
 }
 
-func run(log *log.Logger) error {
+func run(log *zap.SugaredLogger) error {
 
 	// =========================================================================
 	// Configuration
 
-	var cfg struct {
+	cfg := struct {
 		conf.Version
 		Web struct {
-			DebugHost       string        `conf:"default:0.0.0.0:4001"`
-			ReadTimeout     time.Duration `conf:"default:5s"`
-			WriteTimeout    time.Duration `conf:"default:5s"`
-			ShutdownTimeout time.Duration `conf:"default:5s"`
+			DebugHost string `conf:"default:0.0.0.0:4001"`
 		}
 		Expvar struct {
 			Host            string        `conf:"default:0.0.0.0:3001"`
 			Route           string        `conf:"default:/metrics"`
 			ReadTimeout     time.Duration `conf:"default:5s"`
-			WriteTimeout    time.Duration `conf:"default:5s"`
+			WriteTimeout    time.Duration `conf:"default:10s"`
+			IdleTimeout     time.Duration `conf:"default:120s"`
 			ShutdownTimeout time.Duration `conf:"default:5s"`
 		}
 		Collect struct {
-			From string `conf:"default:http://sales-api:4000/debug/vars"`
+			From string `conf:"default:http://localhost:4000/debug/vars"`
 		}
 		Publish struct {
 			To       string        `conf:"default:console"`
 			Interval time.Duration `conf:"default:5s"`
 		}
+	}{
+		Version: conf.Version{
+			SVN:  build,
+			Desc: "copyright information here",
+		},
 	}
-	cfg.Version.SVN = build
-	cfg.Version.Desc = "copyright information here"
 
 	if err := conf.Parse(os.Args[1:], "METRICS", &cfg); err != nil {
 		switch err {
@@ -84,24 +100,30 @@ func run(log *log.Logger) error {
 	if err != nil {
 		return errors.Wrap(err, "generating config for output")
 	}
-	log.Printf("main: Config:\n%v\n", out)
+	log.Infow("startup", "config", out)
 
 	// =========================================================================
-	// Start Debug Service. Not concerned with shutting this down when the
-	// application is being shutdown.
-	//
-	// /debug/pprof - Added to the default mux by the net/http/pprof package.
+	// Start Debug Service
+
+	log.Infow("startup", "status", "debug router started", "host", cfg.Web.DebugHost)
+
+	// The Debug function returns a mux to listen and serve on for all the debug
+	// related endpoints. This include the standard library endpoints.
+
+	debugMux := handlers.DebugStandardLibraryMux()
+
+	// Start the service listening for debug requests.
+	// Not concerned with shutting this down with load shedding.
 	go func() {
-		log.Printf("main: Debug Listening %s", cfg.Web.DebugHost)
-		if err := http.ListenAndServe(cfg.Web.DebugHost, http.DefaultServeMux); err != nil {
-			log.Printf("main: Debug Listener closed : %v", err)
+		if err := http.ListenAndServe(cfg.Web.DebugHost, debugMux); err != nil {
+			log.Errorw("shutdown", "status", "debug router closed", "host", cfg.Web.DebugHost, "ERROR", err)
 		}
 	}()
 
 	// =========================================================================
 	// Start expvar Service
 
-	exp := expvar.New(log, cfg.Expvar.Host, cfg.Expvar.Route, cfg.Expvar.ReadTimeout, cfg.Expvar.WriteTimeout)
+	exp := expvar.New(log, cfg.Expvar.Host, cfg.Expvar.Route, cfg.Expvar.ReadTimeout, cfg.Expvar.WriteTimeout, cfg.Expvar.IdleTimeout)
 	defer exp.Stop(cfg.Expvar.ShutdownTimeout)
 
 	// =========================================================================
@@ -130,9 +152,11 @@ func run(log *log.Logger) error {
 	// Make a channel to listen for an interrupt or terminate signal from the OS.
 	// Use a buffered channel because the signal package requires it.
 	shutdown := make(chan os.Signal, 1)
-	signal.Notify(shutdown, os.Interrupt, syscall.SIGTERM)
+	signal.Notify(shutdown, syscall.SIGINT, syscall.SIGTERM)
 	<-shutdown
 
-	log.Println("main: Start shutdown...")
+	log.Infow("shutdown", "status", "shutdown started")
+	defer log.Infow("shutdown", "status", "shutdown complete")
+
 	return nil
 }
