@@ -1,3 +1,4 @@
+// Package expvar manages the publishing of metrics to stdout.
 package expvar
 
 import (
@@ -9,36 +10,38 @@ import (
 	"time"
 
 	"github.com/dimfeld/httptreemux/v5"
+	"go.uber.org/zap"
 )
 
 // Expvar provide our basic publishing.
 type Expvar struct {
-	log    *log.Logger
+	log    *zap.SugaredLogger
 	server http.Server
 	data   map[string]interface{}
 	mu     sync.Mutex
 }
 
 // New starts a service for consuming the raw expvar stats.
-func New(log *log.Logger, host string, route string, readTimeout, writeTimeout time.Duration) *Expvar {
+func New(log *zap.SugaredLogger, host string, route string, readTimeout, writeTimeout time.Duration, idleTimeout time.Duration) *Expvar {
 	mux := httptreemux.New()
 	exp := Expvar{
 		log: log,
 		server: http.Server{
-			Addr:           host,
-			Handler:        mux,
-			ReadTimeout:    readTimeout,
-			WriteTimeout:   writeTimeout,
-			MaxHeaderBytes: 1 << 20,
+			Addr:         host,
+			Handler:      mux,
+			ReadTimeout:  readTimeout,
+			WriteTimeout: writeTimeout,
+			IdleTimeout:  idleTimeout,
+			ErrorLog:     zap.NewStdLog(log.Desugar()),
 		},
 	}
 
 	mux.Handle("GET", route, exp.handler)
 
 	go func() {
-		log.Println("expvar : API Listening", host)
+		log.Infow("expvar", "status", "API listening", "host", host)
 		if err := exp.server.ListenAndServe(); err != nil {
-			log.Println("expvar : ERROR :", err)
+			log.Errorw("ERROR", zap.Error(err))
 		}
 	}()
 
@@ -47,8 +50,8 @@ func New(log *log.Logger, host string, route string, readTimeout, writeTimeout t
 
 // Stop shuts down the service.
 func (exp *Expvar) Stop(shutdownTimeout time.Duration) {
-	exp.log.Println("expvar : Start shutdown...")
-	defer exp.log.Println("expvar : Completed")
+	exp.log.Infow("expvar", "status", "start shutdown...")
+	defer exp.log.Infow("expvar: Completed")
 
 	// Create context for Shutdown call.
 	ctx, cancel := context.WithTimeout(context.Background(), shutdownTimeout)
@@ -56,9 +59,9 @@ func (exp *Expvar) Stop(shutdownTimeout time.Duration) {
 
 	// Asking listener to shutdown and load shed.
 	if err := exp.server.Shutdown(ctx); err != nil {
-		exp.log.Printf("expvar : Graceful shutdown did not complete in %v : %v", shutdownTimeout, err)
+		exp.log.Errorw("expvar", "status", "graceful shutdown did not complete", "ERROR", err, "shutdownTimeout", shutdownTimeout)
 		if err := exp.server.Close(); err != nil {
-			exp.log.Fatalf("expvar : Could not stop http server: %v", err)
+			exp.log.Errorw("expvar", "status", "could not stop http server", "ERROR", err)
 		}
 	}
 }
@@ -85,7 +88,7 @@ func (exp *Expvar) handler(w http.ResponseWriter, r *http.Request, params map[st
 	exp.mu.Unlock()
 
 	if err := json.NewEncoder(w).Encode(data); err != nil {
-		exp.log.Println("expvar : ERROR :", err)
+		exp.log.Errorw("expvar", "status", "encoding data", "ERROR", err)
 	}
 
 	log.Printf("expvar : (%d) : %s %s -> %s",
