@@ -3,37 +3,25 @@ package product
 
 import (
 	"context"
-	"log"
 	"time"
 
-	"github.com/ardanlabs/service/business/auth"
-	"github.com/ardanlabs/service/business/validate"
+	"github.com/ardanlabs/service/business/sys/auth"
+	"github.com/ardanlabs/service/business/sys/validate"
 	"github.com/ardanlabs/service/foundation/database"
 	"github.com/jmoiron/sqlx"
 	"github.com/pkg/errors"
-	"go.opentelemetry.io/otel/trace"
+	"go.uber.org/zap"
 )
 
-var (
-	// ErrNotFound is used when a specific Product is requested but does not exist.
-	ErrNotFound = errors.New("not found")
-
-	// ErrInvalidID occurs when an ID is not in a valid form.
-	ErrInvalidID = errors.New("ID is not in its proper form")
-
-	// ErrForbidden occurs when a user tries to do something that is forbidden to them according to our access control policies.
-	ErrForbidden = errors.New("attempted action is not allowed")
-)
-
-// Product manages the set of API's for product access.
-type Product struct {
-	log *log.Logger
+// Store manages the set of API's for product access.
+type Store struct {
+	log *zap.SugaredLogger
 	db  *sqlx.DB
 }
 
-// New constructs a Product for api access.
-func New(log *log.Logger, db *sqlx.DB) Product {
-	return Product{
+// NewStore constructs a product store for api access.
+func NewStore(log *zap.SugaredLogger, db *sqlx.DB) Store {
+	return Store{
 		log: log,
 		db:  db,
 	}
@@ -41,22 +29,19 @@ func New(log *log.Logger, db *sqlx.DB) Product {
 
 // Create adds a Product to the database. It returns the created Product with
 // fields like ID and DateCreated populated.
-func (p Product) Create(ctx context.Context, traceID string, claims auth.Claims, np NewProduct, now time.Time) (Info, error) {
-	ctx, span := trace.SpanFromContext(ctx).Tracer().Start(ctx, "business.data.product.create")
-	defer span.End()
-
+func (s Store) Create(ctx context.Context, traceID string, claims auth.Claims, np NewProduct, now time.Time) (Product, error) {
 	if err := validate.Check(np); err != nil {
-		return Info{}, errors.Wrap(err, "validating data")
+		return Product{}, errors.Wrap(err, "validating data")
 	}
 
-	prd := Info{
+	prd := Product{
 		ID:          validate.GenerateID(),
 		Name:        np.Name,
 		Cost:        np.Cost,
 		Quantity:    np.Quantity,
 		UserID:      claims.Subject,
-		DateCreated: now.UTC(),
-		DateUpdated: now.UTC(),
+		DateCreated: now,
+		DateUpdated: now,
 	}
 
 	const q = `
@@ -65,12 +50,8 @@ func (p Product) Create(ctx context.Context, traceID string, claims auth.Claims,
 	VALUES
 		(:product_id, :user_id, :name, :cost, :quantity, :date_created, :date_updated)`
 
-	p.log.Printf("%s: %s: %s", traceID, "product.Create",
-		database.Log(q, prd),
-	)
-
-	if _, err := p.db.NamedExecContext(ctx, q, prd); err != nil {
-		return Info{}, errors.Wrap(err, "inserting product")
+	if err := database.NamedExecContext(ctx, s.log, s.db, traceID, q, prd); err != nil {
+		return Product{}, errors.Wrap(err, "inserting product")
 	}
 
 	return prd, nil
@@ -78,25 +59,22 @@ func (p Product) Create(ctx context.Context, traceID string, claims auth.Claims,
 
 // Update modifies data about a Product. It will error if the specified ID is
 // invalid or does not reference an existing Product.
-func (p Product) Update(ctx context.Context, traceID string, claims auth.Claims, productID string, up UpdateProduct, now time.Time) error {
-	ctx, span := trace.SpanFromContext(ctx).Tracer().Start(ctx, "business.data.product.update")
-	defer span.End()
-
+func (s Store) Update(ctx context.Context, traceID string, claims auth.Claims, productID string, up UpdateProduct, now time.Time) error {
 	if err := validate.CheckID(productID); err != nil {
-		return ErrInvalidID
+		return database.ErrInvalidID
 	}
 	if err := validate.Check(up); err != nil {
 		return errors.Wrap(err, "validating data")
 	}
 
-	prd, err := p.QueryByID(ctx, traceID, productID)
+	prd, err := s.QueryByID(ctx, traceID, productID)
 	if err != nil {
 		return errors.Wrap(err, "updating product")
 	}
 
 	// If you are not an admin and looking to retrieve someone elses product.
 	if !claims.Authorized(auth.RoleAdmin) && prd.UserID != claims.Subject {
-		return ErrForbidden
+		return database.ErrForbidden
 	}
 
 	if up.Name != nil {
@@ -121,11 +99,7 @@ func (p Product) Update(ctx context.Context, traceID string, claims auth.Claims,
 	WHERE
 		product_id = :product_id`
 
-	p.log.Printf("%s: %s: %s", traceID, "product.Update",
-		database.Log(q, prd),
-	)
-
-	if _, err := p.db.NamedExecContext(ctx, q, prd); err != nil {
+	if err := database.NamedExecContext(ctx, s.log, s.db, traceID, q, prd); err != nil {
 		return errors.Wrapf(err, "updating product %s", prd.ID)
 	}
 
@@ -133,17 +107,14 @@ func (p Product) Update(ctx context.Context, traceID string, claims auth.Claims,
 }
 
 // Delete removes the product identified by a given ID.
-func (p Product) Delete(ctx context.Context, traceID string, claims auth.Claims, productID string) error {
-	ctx, span := trace.SpanFromContext(ctx).Tracer().Start(ctx, "business.data.product.delete")
-	defer span.End()
-
+func (s Store) Delete(ctx context.Context, traceID string, claims auth.Claims, productID string) error {
 	if err := validate.CheckID(productID); err != nil {
-		return ErrInvalidID
+		return database.ErrInvalidID
 	}
 
 	// If you are not an admin.
 	if !claims.Authorized(auth.RoleAdmin) {
-		return ErrForbidden
+		return database.ErrForbidden
 	}
 
 	data := struct {
@@ -158,11 +129,7 @@ func (p Product) Delete(ctx context.Context, traceID string, claims auth.Claims,
 	WHERE
 		product_id = :product_id`
 
-	p.log.Printf("%s: %s: %s", traceID, "product.Delete",
-		database.Log(q, data),
-	)
-
-	if _, err := p.db.NamedExecContext(ctx, q, data); err != nil {
+	if err := database.NamedExecContext(ctx, s.log, s.db, traceID, q, data); err != nil {
 		return errors.Wrapf(err, "deleting product %s", data.ProductID)
 	}
 
@@ -170,10 +137,7 @@ func (p Product) Delete(ctx context.Context, traceID string, claims auth.Claims,
 }
 
 // Query gets all Products from the database.
-func (p Product) Query(ctx context.Context, traceID string, pageNumber int, rowsPerPage int) ([]Info, error) {
-	ctx, span := trace.SpanFromContext(ctx).Tracer().Start(ctx, "business.data.product.query")
-	defer span.End()
-
+func (s Store) Query(ctx context.Context, traceID string, pageNumber int, rowsPerPage int) ([]Product, error) {
 	data := struct {
 		Offset      int `db:"offset"`
 		RowsPerPage int `db:"rows_per_page"`
@@ -197,14 +161,10 @@ func (p Product) Query(ctx context.Context, traceID string, pageNumber int, rows
 		user_id
 	OFFSET :offset ROWS FETCH NEXT :rows_per_page ROWS ONLY`
 
-	p.log.Printf("%s: %s: %s", traceID, "product.Query",
-		database.Log(q, data),
-	)
-
-	var products []Info
-	if err := database.NamedQuerySlice(ctx, p.db, q, data, &products); err != nil {
+	var products []Product
+	if err := database.NamedQuerySlice(ctx, s.log, s.db, traceID, q, data, &products); err != nil {
 		if err == database.ErrNotFound {
-			return nil, ErrNotFound
+			return nil, database.ErrNotFound
 		}
 		return nil, errors.Wrap(err, "selecting products")
 	}
@@ -213,12 +173,9 @@ func (p Product) Query(ctx context.Context, traceID string, pageNumber int, rows
 }
 
 // QueryByID finds the product identified by a given ID.
-func (p Product) QueryByID(ctx context.Context, traceID string, productID string) (Info, error) {
-	ctx, span := trace.SpanFromContext(ctx).Tracer().Start(ctx, "business.data.product.querybyid")
-	defer span.End()
-
+func (s Store) QueryByID(ctx context.Context, traceID string, productID string) (Product, error) {
 	if err := validate.CheckID(productID); err != nil {
-		return Info{}, ErrInvalidID
+		return Product{}, database.ErrInvalidID
 	}
 
 	data := struct {
@@ -241,16 +198,12 @@ func (p Product) QueryByID(ctx context.Context, traceID string, productID string
 	GROUP BY
 		p.product_id`
 
-	p.log.Printf("%s: %s: %s", traceID, "product.QueryByID",
-		database.Log(q, data),
-	)
-
-	var prd Info
-	if err := database.NamedQueryStruct(ctx, p.db, q, data, &prd); err != nil {
+	var prd Product
+	if err := database.NamedQueryStruct(ctx, s.log, s.db, traceID, q, data, &prd); err != nil {
 		if err == database.ErrNotFound {
-			return Info{}, ErrNotFound
+			return Product{}, database.ErrNotFound
 		}
-		return Info{}, errors.Wrapf(err, "selecting user %q", data.ProductID)
+		return Product{}, errors.Wrapf(err, "selecting product %q", data.ProductID)
 	}
 
 	return prd, nil
