@@ -2,14 +2,13 @@
 package schema
 
 import (
-	_ "embed" // Embed all sql documents
+	"context"
+	_ "embed" // Calls init function.
 
-	"bufio"
-	"strconv"
-	"strings"
-
-	"github.com/dimiro1/darwin"
+	"github.com/ardanlabs/darwin"
+	"github.com/ardanlabs/service/foundation/database"
 	"github.com/jmoiron/sqlx"
+	"github.com/pkg/errors"
 )
 
 var (
@@ -25,15 +24,27 @@ var (
 
 // Migrate attempts to bring the schema for db up to date with the migrations
 // defined in this package.
-func Migrate(db *sqlx.DB) error {
-	driver := darwin.NewGenericDriver(db.DB, darwin.PostgresDialect{})
-	d := darwin.New(driver, parseMigrations(schemaDoc), nil)
+func Migrate(ctx context.Context, db *sqlx.DB) error {
+	if err := database.StatusCheck(ctx, db); err != nil {
+		return errors.Wrap(err, "status check database")
+	}
+
+	driver, err := darwin.NewGenericDriver(db.DB, darwin.PostgresDialect{})
+	if err != nil {
+		return errors.Wrap(err, "construct darwin driver")
+	}
+
+	d := darwin.New(driver, darwin.ParseMigrations(schemaDoc))
 	return d.Migrate()
 }
 
 // Seed runs the set of seed-data queries against db. The queries are ran in a
 // transaction and rolled back if any fail.
-func Seed(db *sqlx.DB) error {
+func Seed(ctx context.Context, db *sqlx.DB) error {
+	if err := database.StatusCheck(ctx, db); err != nil {
+		return errors.Wrap(err, "status check database")
+	}
+
 	tx, err := db.Begin()
 	if err != nil {
 		return err
@@ -65,42 +76,4 @@ func DeleteAll(db *sqlx.DB) error {
 	}
 
 	return tx.Commit()
-}
-
-func parseMigrations(s string) []darwin.Migration {
-	var migs []darwin.Migration
-
-	scanner := bufio.NewScanner(strings.NewReader(s))
-	scanner.Split(bufio.ScanLines)
-
-	var mig darwin.Migration
-	var script string
-	for scanner.Scan() {
-		v := strings.ToLower(scanner.Text())
-		switch {
-		case len(v) >= 5 && (v[:6] == "-- ver" || v[:5] == "--ver"):
-			mig.Script = script
-			migs = append(migs, mig)
-
-			mig = darwin.Migration{}
-			script = ""
-
-			f, err := strconv.ParseFloat(strings.TrimSpace(v[11:]), 64)
-			if err != nil {
-				return nil
-			}
-			mig.Version = f
-
-		case len(v) >= 5 && (v[:6] == "-- des" || v[:5] == "--des"):
-			mig.Description = strings.TrimSpace(v[15:])
-
-		default:
-			script += v + "\n"
-		}
-	}
-
-	mig.Script = script
-	migs = append(migs, mig)
-
-	return migs[1:]
 }
