@@ -2,31 +2,38 @@
 package database
 
 import (
-	_ "github.com/lib/pq" // Initialize the database driver.
-
 	"context"
 	"errors"
 	"fmt"
 	"net/url"
 	"reflect"
 	"strings"
+	"time"
 
 	"github.com/jmoiron/sqlx"
-	"go.opentelemetry.io/otel/trace"
+	_ "github.com/lib/pq" // Calls init function.
+	"go.opentelemetry.io/otel"
+	"go.opentelemetry.io/otel/attribute"
+	"go.uber.org/zap"
 )
 
+// Set of error variables for CRUD operations.
 var (
-	// ErrNotFound is used when a query is performed with no data.
-	ErrNotFound = errors.New("not found")
+	ErrNotFound              = errors.New("not found")
+	ErrInvalidID             = errors.New("ID is not in its proper form")
+	ErrAuthenticationFailure = errors.New("authentication failed")
+	ErrForbidden             = errors.New("attempted action is not allowed")
 )
 
 // Config is the required properties to use the database.
 type Config struct {
-	User       string
-	Password   string
-	Host       string
-	Name       string
-	DisableTLS bool
+	User         string
+	Password     string
+	Host         string
+	Name         string
+	MaxIdleConns int
+	MaxOpenConns int
+	DisableTLS   bool
 }
 
 // Open knows how to open a database connection based on the configuration.
@@ -48,20 +55,80 @@ func Open(cfg Config) (*sqlx.DB, error) {
 		RawQuery: q.Encode(),
 	}
 
-	return sqlx.Open("postgres", u.String())
+	db, err := sqlx.Open("postgres", u.String())
+	if err != nil {
+		return nil, err
+	}
+	db.SetMaxIdleConns(cfg.MaxIdleConns)
+	db.SetMaxOpenConns(cfg.MaxOpenConns)
+
+	return db, nil
+}
+
+// StatusCheck returns nil if it can successfully talk to the database. It
+// returns a non-nil error otherwise.
+func StatusCheck(ctx context.Context, db *sqlx.DB) error {
+
+	// First check we can ping the database.
+	var pingError error
+	for attempts := 1; ; attempts++ {
+		pingError = db.Ping()
+		if pingError == nil {
+			break
+		}
+		time.Sleep(time.Duration(attempts) * 100 * time.Millisecond)
+		if ctx.Err() != nil {
+			return ctx.Err()
+		}
+	}
+
+	// Make sure we didn't timeout or be cancelled.
+	if ctx.Err() != nil {
+		return ctx.Err()
+	}
+
+	// Run a simple query to determine connectivity. Running this query forces a
+	// round trip through the database.
+	const q = `SELECT true`
+	var tmp bool
+	return db.QueryRowContext(ctx, q).Scan(&tmp)
+}
+
+// NamedExecContext is a helper function to execute a CUD operation with
+// logging and tracing.
+func NamedExecContext(ctx context.Context, log *zap.SugaredLogger, db *sqlx.DB, traceID string, query string, data interface{}) error {
+	q := queryString(query, data)
+	log.Infow("database.NamedExecContext", "traceid", traceID, "query", q)
+
+	ctx, span := otel.GetTracerProvider().Tracer("").Start(ctx, "database.query")
+	span.SetAttributes(attribute.String("query", q))
+	defer span.End()
+
+	if _, err := db.NamedExecContext(ctx, query, data); err != nil {
+		return err
+	}
+
+	return nil
 }
 
 // NamedQuerySlice is a helper function for executing queries that return a
 // collection of data to be unmarshaled into a slice.
-func NamedQuerySlice(ctx context.Context, db *sqlx.DB, query string, data interface{}, dest interface{}) error {
-	rows, err := db.NamedQueryContext(ctx, query, data)
-	if err != nil {
-		return err
-	}
+func NamedQuerySlice(ctx context.Context, log *zap.SugaredLogger, db *sqlx.DB, traceID string, query string, data interface{}, dest interface{}) error {
+	q := queryString(query, data)
+	log.Infow("database.NamedQuerySlice", "traceid", traceID, "query", q)
+
+	ctx, span := otel.GetTracerProvider().Tracer("").Start(ctx, "database.query")
+	span.SetAttributes(attribute.String("query", q))
+	defer span.End()
 
 	val := reflect.ValueOf(dest)
 	if val.Kind() != reflect.Ptr || val.Elem().Kind() != reflect.Slice {
 		return errors.New("must provide a pointer to a slice")
+	}
+
+	rows, err := db.NamedQueryContext(ctx, query, data)
+	if err != nil {
+		return err
 	}
 
 	slice := val.Elem()
@@ -78,7 +145,14 @@ func NamedQuerySlice(ctx context.Context, db *sqlx.DB, query string, data interf
 
 // NamedQueryStruct is a helper function for executing queries that return a
 // single value to be unmarshalled into a struct type.
-func NamedQueryStruct(ctx context.Context, db *sqlx.DB, query string, data interface{}, dest interface{}) error {
+func NamedQueryStruct(ctx context.Context, log *zap.SugaredLogger, db *sqlx.DB, traceID string, query string, data interface{}, dest interface{}) error {
+	q := queryString(query, data)
+	log.Infow("database.NamedQuerySlice", "traceid", traceID, "query", q)
+
+	ctx, span := otel.GetTracerProvider().Tracer("").Start(ctx, "database.query")
+	span.SetAttributes(attribute.String("query", q))
+	defer span.End()
+
 	rows, err := db.NamedQueryContext(ctx, query, data)
 	if err != nil {
 		return err
@@ -94,23 +168,8 @@ func NamedQueryStruct(ctx context.Context, db *sqlx.DB, query string, data inter
 	return nil
 }
 
-// StatusCheck returns nil if it can successfully talk to the database. It
-// returns a non-nil error otherwise.
-func StatusCheck(ctx context.Context, db *sqlx.DB) error {
-	ctx, span := trace.SpanFromContext(ctx).Tracer().Start(ctx, "foundation.database.statuscheck")
-	defer span.End()
-
-	// Run a simple query to determine connectivity. The db has a "Ping" method
-	// but it can false-positive when it was previously able to talk to the
-	// database but the database has since gone away. Running this query forces a
-	// round trip to the database.
-	const q = `SELECT true`
-	var tmp bool
-	return db.QueryRowContext(ctx, q).Scan(&tmp)
-}
-
-// Log provides a pretty print version of the query and parameters.
-func Log(query string, args ...interface{}) string {
+// queryString provides a pretty print version of the query and parameters.
+func queryString(query string, args ...interface{}) string {
 	query, params, err := sqlx.Named(query, args)
 	if err != nil {
 		return err.Error()
@@ -132,5 +191,5 @@ func Log(query string, args ...interface{}) string {
 	query = strings.Replace(query, "\t", "", -1)
 	query = strings.Replace(query, "\n", " ", -1)
 
-	return fmt.Sprintf("[%s]\n", strings.Trim(query, " "))
+	return strings.Trim(query, " ")
 }
