@@ -30,13 +30,6 @@ type Values struct {
 // framework.
 type Handler func(ctx context.Context, w http.ResponseWriter, r *http.Request) error
 
-// registered keeps track of handlers registered to the http default server
-// mux. This is a singleton and used by the standard library for metrics
-// and profiling. The application may want to add other handlers like
-// readiness and liveness to that mux. If this is not tracked, the routes
-// could try to be registered more than once, causing a panic.
-var registered = make(map[string]bool)
-
 // App is the entrypoint into our application and what configures our context
 // object for each of our http handlers. Feel free to add any configuration
 // data/logic on this App struct.
@@ -81,29 +74,9 @@ func (a *App) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	a.otmux.ServeHTTP(w, r)
 }
 
-// HandleDebug sets a handler function for a given HTTP method and path pair
-// to the default http package server mux. /debug is added to the path.
-func (a *App) HandleDebug(method string, path string, handler Handler, mw ...Middleware) {
-	a.handle(true, method, path, handler, mw...)
-}
-
 // Handle sets a handler function for a given HTTP method and path pair
 // to the application server mux.
 func (a *App) Handle(method string, path string, handler Handler, mw ...Middleware) {
-	a.handle(false, method, path, handler, mw...)
-}
-
-// handle performs the real work of applying boilerplate and framework code
-// for a handler.
-func (a *App) handle(debug bool, method string, path string, handler Handler, mw ...Middleware) {
-	if debug {
-		// Track all the handlers that are being registered so we don't have
-		// the same handlers registered twice to this singleton.
-		if _, exists := registered[method+path]; exists {
-			return
-		}
-		registered[method+path] = true
-	}
 
 	// First wrap handler specific middleware around this handler.
 	handler = wrapMiddleware(mw, handler)
@@ -114,15 +87,17 @@ func (a *App) handle(debug bool, method string, path string, handler Handler, mw
 	// The function to execute for each request.
 	h := func(w http.ResponseWriter, r *http.Request) {
 
-		// Start or expand a distributed trace.
+		// Pull the context from the request and
+		// use it as a separate parameter.
 		ctx := r.Context()
-		ctx, span := trace.SpanFromContext(ctx).Tracer().Start(ctx, r.URL.Path)
-		defer span.End()
+
+		// Capture the parent request span from the context.
+		span := trace.SpanFromContext(ctx)
 
 		// Set the context with the required values to
 		// process the request.
 		v := Values{
-			TraceID: span.SpanContext().TraceID.String(),
+			TraceID: span.SpanContext().TraceID().String(),
 			Now:     time.Now(),
 		}
 		ctx = context.WithValue(ctx, KeyValues, &v)
@@ -134,18 +109,5 @@ func (a *App) handle(debug bool, method string, path string, handler Handler, mw
 		}
 	}
 
-	// Add this handler for the specified verb and route.
-	if debug {
-		f := func(w http.ResponseWriter, r *http.Request) {
-			switch {
-			case r.Method == method:
-				h(w, r)
-			default:
-				w.WriteHeader(http.StatusNotFound)
-			}
-		}
-		http.DefaultServeMux.HandleFunc("/debug"+path, f)
-		return
-	}
 	a.mux.Handle(method, path, h)
 }
